Bug fixes.
I found that sometimes with some of my accounts being in the same servers as one another that the printed output would get scrambled together sometimes. So I've added do while to most of our messages recieved printing outputs. Ontop of that while I was in the middle of adding the aformentioned fix a friend reported an unhandled data return to me that I have added a case handle for.